LayoutTests/inspector/indexeddb/requestDatabaseNames-expected.txt

CONSOLE MESSAGE: line 10: Created Database 'Database1'
CONSOLE MESSAGE: line 10: Created Database 'Database2'

PASS: No IndexedDB databases should exist initially
Created Database 'Database1'
PASS: A single IndexedDB database should exist
["Database1"]
Created Database 'Database2'
PASS: Two IndexedDB databases should exist
["Database1","Database2"]
